DISCO-3181 Rename GCP util to GCS
