+++ tests/unit/content_handler/test_gcp_uploader.py
@@ -1,11 +1,11 @@
-"""Unit tests for merino.gcs.gcp_uploader"""
+"""Unit tests for merino.gcs.gcs_uploader"""
 
 from logging import ERROR, INFO, LogRecord
 
 import pytest
 from pytest import LogCaptureFixture
 
-from merino.utils.gcs.gcp_uploader import GcsUploader
+from merino.utils.gcs.gcs_uploader import GcsUploader
 from merino.utils.gcs.models import Image
 from tests.types import FilterCaplogFixture
 
@@ -13,7 +13,7 @@ from tests.types import FilterCaplogFixture
 @pytest.fixture
 def mock_gcs_client(mocker, mock_gcs_bucket):
     """Return a mock GCS Client instance"""
-    mock_client = mocker.patch("merino.utils.gcs.gcp_uploader.Client").return_value
+    mock_client = mocker.patch("merino.utils.gcs.gcs_uploader.Client").return_value
     mock_client.get_bucket.return_value = mock_gcs_bucket
     return mock_client
 
@@ -21,13 +21,13 @@ def mock_gcs_client(mocker, mock_gcs_bucket):
 @pytest.fixture
 def mock_gcs_blob(mocker):
     """Return a mock GCS Blob instance"""
-    return mocker.patch("merino.utils.gcs.gcp_uploader.Blob").return_value
+    return mocker.patch("merino.utils.gcs.gcs_uploader.Blob").return_value
 
 
 @pytest.fixture
 def mock_most_recent_gcs_blob(mocker):
     """Return a mock GCS Blob instance"""
-    most_recent_blob = mocker.patch("merino.utils.gcs.gcp_uploader.Blob").return_value
+    most_recent_blob = mocker.patch("merino.utils.gcs.gcs_uploader.Blob").return_value
     most_recent_blob.name = "20220101120555_top_picks.json"
     return most_recent_blob
 
@@ -35,7 +35,7 @@ def mock_most_recent_gcs_blob(mocker):
 @pytest.fixture
 def mock_gcs_bucket(mocker):
     """Return a mock GCS Bucket instance"""
-    return mocker.patch("merino.utils.gcs.gcp_uploader.Bucket").return_value
+    return mocker.patch("merino.utils.gcs.gcs_uploader.Bucket").return_value
 
 
 @pytest.fixture
@@ -84,13 +84,13 @@ def test_upload_image_with_non_https_cdn_host_name(
     caplog.set_level(INFO)
 
     # creating the uploader object with cdn host name not containing "https"
-    gcp_uploader = GcsUploader(mock_gcs_client, test_bucket_name, test_cdn_host_name)
+    gcs_uploader = GcsUploader(mock_gcs_client, test_bucket_name, test_cdn_host_name)
 
     # force upload is set to FALSE by default
-    result = gcp_uploader.upload_image(test_image, test_destination_name)
+    result = gcs_uploader.upload_image(test_image, test_destination_name)
 
     # capture logger info output
-    log_records: list[LogRecord] = filter_caplog(caplog.records, "merino.utils.gcs.gcp_uploader")
+    log_records: list[LogRecord] = filter_caplog(caplog.records, "merino.utils.gcs.gcs_uploader")
 
     assert result == f"https://{test_cdn_host_name}/{test_destination_name}"
     # assert on logger calls
@@ -112,13 +112,13 @@ def test_upload_image_with_https_cdn_host_name(
     caplog.set_level(INFO)
 
     # creating the uploader object with cdn host name containing "https"
-    gcp_uploader = GcsUploader(mock_gcs_client, test_bucket_name, test_https_cdn_host_name)
+    gcs_uploader = GcsUploader(mock_gcs_client, test_bucket_name, test_https_cdn_host_name)
 
     # force upload is set to FALSE by default
-    result = gcp_uploader.upload_image(test_image, test_destination_name)
+    result = gcs_uploader.upload_image(test_image, test_destination_name)
 
     # capture logger info output
-    log_records: list[LogRecord] = filter_caplog(caplog.records, "merino.utils.gcs.gcp_uploader")
+    log_records: list[LogRecord] = filter_caplog(caplog.records, "merino.utils.gcs.gcs_uploader")
 
     assert result == f"{test_https_cdn_host_name}/{test_destination_name}"
     # assert on logger calls
@@ -139,10 +139,10 @@ def test_get_most_recent_file_with_two_files(
     # set the mock bucket's blob list to the two mocked blobs
     mock_gcs_bucket.list_blobs.return_value = [mock_most_recent_gcs_blob, mock_gcs_blob]
 
-    gcp_uploader = GcsUploader(mock_gcs_client, test_bucket_name, test_https_cdn_host_name)
+    gcs_uploader = GcsUploader(mock_gcs_client, test_bucket_name, test_https_cdn_host_name)
 
     excluded_file: str = "excluded.json"
-    result = gcp_uploader.get_most_recent_file(
+    result = gcs_uploader.get_most_recent_file(
         exclusion=excluded_file, sort_key=lambda blob: blob.name
     )
 
@@ -167,10 +167,10 @@ def test_get_most_recent_file_with_excluded_file(
     # bucket only contains the excluded file
     mock_gcs_bucket.list_blobs.return_value = [mock_gcs_blob]
 
-    gcp_uploader = GcsUploader(mock_gcs_client, test_bucket_name, test_https_cdn_host_name)
+    gcs_uploader = GcsUploader(mock_gcs_client, test_bucket_name, test_https_cdn_host_name)
 
     # call the method with the exclusion argument set to the excluded file
-    result = gcp_uploader.get_most_recent_file(
+    result = gcs_uploader.get_most_recent_file(
         exclusion=excluded_file, sort_key=lambda blob: blob.name
     )
 
@@ -195,14 +195,14 @@ def test_upload_content_with_forced_upload_false_and_existing_blob(
     mock_gcs_client.bucket.return_value = mock_gcs_bucket
     mock_gcs_bucket.blob.return_value = mock_gcs_blob
 
-    gcp_uploader = GcsUploader(mock_gcs_client, test_bucket_name, test_https_cdn_host_name)
+    gcs_uploader = GcsUploader(mock_gcs_client, test_bucket_name, test_https_cdn_host_name)
     content = bytes(255)
 
     # call the method
-    result = gcp_uploader.upload_content(content, test_destination_name)
+    result = gcs_uploader.upload_content(content, test_destination_name)
 
     # capture logger info output
-    log_records: list[LogRecord] = filter_caplog(caplog.records, "merino.utils.gcs.gcp_uploader")
+    log_records: list[LogRecord] = filter_caplog(caplog.records, "merino.utils.gcs.gcs_uploader")
 
     assert result == mock_gcs_blob
     assert len(log_records) == 0
@@ -228,13 +228,13 @@ def test_upload_content_with_forced_upload_true_and_existing_blob(
     mock_gcs_client.bucket.return_value = mock_gcs_bucket
     mock_gcs_bucket.blob.return_value = mock_gcs_blob
 
-    gcp_uploader = GcsUploader(mock_gcs_client, test_bucket_name, test_https_cdn_host_name)
+    gcs_uploader = GcsUploader(mock_gcs_client, test_bucket_name, test_https_cdn_host_name)
     content = bytes(255)
 
-    result = gcp_uploader.upload_content(content, test_destination_name, forced_upload=True)
+    result = gcs_uploader.upload_content(content, test_destination_name, forced_upload=True)
 
     # capture logger info output
-    log_records: list[LogRecord] = filter_caplog(caplog.records, "merino.utils.gcs.gcp_uploader")
+    log_records: list[LogRecord] = filter_caplog(caplog.records, "merino.utils.gcs.gcs_uploader")
 
     assert result == mock_gcs_blob
     assert len(log_records) == 1
@@ -266,13 +266,13 @@ def test_upload_content_with_exception_thrown(
     # make the blob.make_public() method throw a run time exception
     mock_gcs_blob.make_public.side_effect = RuntimeError("test-exception")
 
-    gcp_uploader = GcsUploader(mock_gcs_client, test_bucket_name, test_https_cdn_host_name)
+    gcs_uploader = GcsUploader(mock_gcs_client, test_bucket_name, test_https_cdn_host_name)
 
     # call the method
-    gcp_uploader.upload_content(content, test_destination_name, forced_upload=True)
+    gcs_uploader.upload_content(content, test_destination_name, forced_upload=True)
 
     # capture logger error output
-    log_records: list[LogRecord] = filter_caplog(caplog.records, "merino.utils.gcs.gcp_uploader")
+    log_records: list[LogRecord] = filter_caplog(caplog.records, "merino.utils.gcs.gcs_uploader")
 
     assert len(log_records) == 1
     assert log_records[0].message.startswith(
