+++ tests/unit/jobs/navigational_suggestions/test_domain_metadata_uploader.py
@@ -182,13 +182,13 @@ def fixture_json_domain_data_latest() -> str:
 @pytest.fixture
 def mock_gcs_client(mocker):
     """Return a mock GCS Client instance"""
-    return mocker.patch("merino.utils.gcs.gcp_uploader.Client").return_value
+    return mocker.patch("merino.utils.gcs.gcs_uploader.Client").return_value
 
 
 @pytest.fixture
 def mock_gcs_blob(mocker):
     """Return a mock GCS Bucket instance"""
-    return mocker.patch("merino.utils.gcs.gcp_uploader.Blob").return_value
+    return mocker.patch("merino.utils.gcs.gcs_uploader.Blob").return_value
 
 
 @pytest.fixture(name="remote_blob", autouse=True)
